+++ src/mass_driver/drivers/__init__.py
@@ -2,5 +2,6 @@
 from mass_driver.drivers.counter import Counter
 from mass_driver.drivers.jsonpatch import JsonPatch
 from mass_driver.drivers.poetry import Poetry
+from mass_driver.drivers.precommit import PrecommitAutoupdate
 
-__all__ = ["Counter", "Poetry", "JsonPatch"]
+__all__ = ["Counter", "Poetry", "JsonPatch", "PrecommitAutoupdate"]